Changed damage calculating code

+++ src/main/java/com/github/tcgeneric/wargame/events/EntityDamagedEvent.kt
@@ -5,7 +5,7 @@ import com.github.tcgeneric.wargame.entity.units.Unit
 import org.bukkit.event.Event
 import org.bukkit.event.HandlerList
 
-class EntityDamagedEvent(val unit:Entity, val damage:Int):Event() {
+class EntityDamagedEvent(val unit:Entity, val damage:Double):Event() {
     private val handlerList:HandlerList = HandlerList()
     override fun getHandlers(): HandlerList {
         return handlerList
